requirements.txt
# AsyncIO
aiohttp<4,>=3.7.3
aiostream==0.4.3

# HTML crawling & parsing
beautifulsoup4==4.6.0
html5lib==1.1

# Twitter streaming & authentication
tweepy==4.1.0
oauthlib<4,>=3.1.0

# Miscellaneous
colorama==0.3.9
pyyaml=5.3.1